backend/requirements.txt
crawl4ai>=0.2.77
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
openai>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0